MCP/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
